refactor: Move replication function on its own file :bowling:
